moved accelerators
