+++ scripts/utils/facades/TestRunner.ts
@@ -1,9 +1,13 @@
-import { Account, Provider } from "starknet";
+import { Account, Provider, TypedContractV2 } from "starknet";
 import { ERC20Facade } from "./erc20Facade";
 import { VaultFacade } from "./vaultFacade";
 import { ApprovalArgs, Constants, DepositArgs, WithdrawArgs } from "./types";
-import { getCustomAccount } from "../helpers/common";
+import { getAccount, getCustomAccount, stringToHex } from "../helpers/common";
 import { liquidityProviders, optionBidders } from "../constants";
+import { MarketAggregatorFacade } from "./marketAggregatorFacade";
+import { getOptionRoundContract } from "../helpers/setup";
+import { getNow, timeskipNextBlock } from "../katana";
+import { vaultABI } from "../../abi";
 
 export type ResultSheet = {
   accounts: Array<Account>;
@@ -153,6 +157,116 @@ export class TestRunner {
     for (const withdrawArgs of withdrawData) {
       await this.vaultFacade.withdraw(withdrawArgs);
     }
+  }
+
+   //@note Only works for katana dev instance with a --dev flag
+   async startAuctionBystander(provider: Provider, constants: Constants) {
+    const devAccount = getAccount("dev", provider);
+    //Set market aggregator reserve_price
+    const marketAggregatorString =
+      await this.vaultFacade.vaultContract.get_market_aggregator();
+    const marketAggregatorAddress = "0x" + stringToHex(marketAggregatorString);
+    const marketAggFacade = new MarketAggregatorFacade(
+      marketAggregatorAddress,
+      provider
+    );
+    const optionRound = await getOptionRoundContract(
+      provider,
+      this.vaultFacade.vaultContract
+    );
+    const startDate = await optionRound.get_auction_start_date();
+    const settleDate = await optionRound.get_option_settlement_date();
+    await marketAggFacade.setReservePrice(
+      devAccount,
+      startDate,
+      settleDate,
+      constants.reservePrice
+    );
+    await marketAggFacade.setCapLevel(
+      devAccount,
+      startDate,
+      settleDate,
+      constants.capLevel
+    );
+
+    await marketAggFacade.setStrikePrice(
+      devAccount,
+      startDate,
+      settleDate,
+      constants.strikePrice
+    );
+    await marketAggFacade.setTWAP(
+      devAccount,
+      startDate,
+      settleDate,
+      constants.settlementPrice
+    );
+    await this.vaultFacade.vaultContract.update_round_params();
+
+    await this.accelerateToAuctioning();
+
+    this.vaultFacade.vaultContract.connect(devAccount);
+    await this.vaultFacade.vaultContract.start_auction();
+  }
+
+  async accelerateToAuctioning() {
+    const optionRoundContract = await getOptionRoundContract(
+      this.provider,
+      this.vaultFacade.vaultContract
+    );
+    const currentTime = await getNow(this.provider);
+    const auctionStartDate = await optionRoundContract.get_auction_start_date();
+  
+    console.log("currentTime:",currentTime,"\nauctionStartDate:",auctionStartDate);;
+    await timeskipNextBlock(
+      Number(auctionStartDate) - Number(currentTime),
+      this.provider.channel.nodeUrl
+    );
+  }
+
+  async endAuctionBystander(provider: Provider) {
+    const devAccount = getAccount("dev", provider);
+    await this.accelerateToRunning();
+    this.vaultFacade.vaultContract.connect(devAccount);
+    await this.vaultFacade.vaultContract.end_auction();
+  }
+
+  async accelerateToRunning() {
+    const optionRoundContract = await getOptionRoundContract(
+      this.provider,
+      this.vaultFacade.vaultContract
+    );
+  
+    const currentTime = await getNow(this.provider);
+    const auctionEndDate = await optionRoundContract.get_auction_end_date();
+  
+    await timeskipNextBlock(
+      Number(auctionEndDate) - Number(currentTime),
+      this.provider.channel.nodeUrl
+    );
+  }
+
+  async settleOptionRoundBystander(provider: Provider) {
+    await this.accelerateToSettled();
+    const devAccount = getAccount("dev", provider);
+    await this.vaultFacade.settleOptionRound(devAccount);
+  }
+
+  async  accelerateToSettled( ) {
+    const optionRoundContract = await getOptionRoundContract(
+      this.provider,
+      this.vaultFacade.vaultContract
+    );
+  
+    const currentTime = await getNow(this.provider);
+    const optionSettleDate =
+      await optionRoundContract.get_option_settlement_date();
+  
+    await timeskipNextBlock(
+      Number(optionSettleDate) - Number(currentTime)+1,
+      this.provider.channel.nodeUrl
+    );
+   
   }
 }
